backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.2
rapidfuzz==3.5.2
openai==1.3.5
python-dotenv==1.0.0
